Add a test framework for testing Volos on an Apigee server. Deploy a proxy to the server and then run the tests remotely via HTTP.
