+++ package.json
@@ -18,7 +18,8 @@
     "commander": "2.2.x",
     "inquirer": "0.5.x",
     "apigee-access": "1.0.x",
-    "debug": "1.0.x"
+    "debug": "1.0.x",
+    "apigeetool": "x.x.x"
   },
   "scripts": {
     "install": "bin/link",
