+++ serversidetests/proxybase/apiproxy/proxies/apigeecache.xml
@@ -1,0 +1,8 @@
+<ProxyEndpoint name="apigeecache">
+  <HTTPProxyConnection>
+    <BasePath>/volostests/apigeecache</BasePath>
+  </HTTPProxyConnection>
+  <RouteRule name="default">
+    <TargetEndpoint>apigeecache</TargetEndpoint>
+  </RouteRule>
+</ProxyEndpoint>
