+++ serversidetests/proxybase/apiproxy/targets/apigeeoauth.xml
@@ -1,0 +1,5 @@
+<TargetEndpoint name="apigeeoauth">
+  <ScriptTarget>
+    <ResourceURL>node://apigeeoauth.js</ResourceURL>
+  </ScriptTarget>
+</TargetEndpoint>
