+++ serversidetests/proxybase/apiproxy/volostest.xml
@@ -1,0 +1,1 @@
+<APIProxy name="volostest"/>
